kcompressd: Fix OOB array access from round-robin counter overflow
kcompressd now distributes bios across kcompressd workers in a round-robin fashion
using next_kcompressd_idx.

next_kcompressd_idx is an atomic_t, a signed 32-bit integer. When it wraps and
becomes negative, a signed modulo in C keeps the sign of the dividend, producing a
negative remainder. Using this negative value as an index causes an out-of-bounds
access.

Make the round-robin indices unsigned int, since they should never be negative in
the first place. Also, cast the dividend to unsigned *before* the modulo operation,
ensuring the result always falls within [0, local_nr - 1]. Thus, we stay in valid
bounds.

Fixes: 2bc4e8172c5d ("kcompressd: Fix load balancing in bio distribution")

+++ drivers/block/zram/kcompressd.c
@@ -280,7 +280,7 @@ static atomic_t next_kcompressd_idx = ATOMIC_INIT(0);
 
 int schedule_bio_write(void *mem, struct bio *bio, compress_callback cb)
 {
-	int i, start_idx, idx;
+	unsigned int i, start_idx, idx;
 	bool submit_success = false;
 	size_t sz_work = sizeof(struct write_work);
 
@@ -298,7 +298,7 @@ int schedule_bio_write(void *mem, struct bio *bio, compress_callback cb)
 	if (!local_nr || !current_is_kswapd())
 		return -EBUSY;
 
-	start_idx = atomic_fetch_inc(&next_kcompressd_idx) % local_nr;
+	start_idx = (unsigned int)atomic_fetch_inc(&next_kcompressd_idx) % local_nr;
 
 	for (i = 0; i < local_nr; i++) {
 		idx = (start_idx + i) % local_nr;
